requirements.txt
ipywidgets
dvc==3.51.2
dvc-gdrive==3.0.1
hyperopt==0.2.7
mlflow==2.14.2
catboost==1.2.5
python-dotenv==1.0.1
fastapi==0.112.0
pandas==2.2.2